+++ CHANGELOG.rst
@@ -7,7 +7,9 @@ Version 0.2.1
 
 - Part of fix for https://github.com/paul-ollis/vpe_syntax/issues/7.
 
-  Changes added in version 0.2.1 of VPE-Sitter also required.
+- Fix for the colour scheme tweaker.
+
+  The NONE colour toggle ("N" key) did not work.
 
 
 Version 0.2.0